[env](compile)import compile_check_avoid to prevent some code from compile-checked. (#47862)
### What problem does this PR solve?

1.
This code uses templates, and errors like the following are likely to
occur, mainly due to literal type mismatches:
be/src/vec/common/cow.h:409:39: warning: implicit conversion loses
integer precision: 'int' to 'value_type' (aka 'unsigned char')
[-Wimplicit-int-conversion]
409 | return MutablePtr(new Derived(std::forward<Args>(args)...));
| ~~~~~~~ ^~~~~~~~~~~~~~~~~~~~~~~~
    ColumnPtr res_data_column = ColumnUInt8::create(1, 1);
2. 
Some third-party libraries or files generated in the gensrc directory
may have some compile check failures.
3.
Also, some compile check errors have been fixed.

Problem Summary:

### Release note

None

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [ ] Manual test (add detailed scripts or steps below)
    - [x] No need to test or manual test. Explain why:
- [x] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [x] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [x] No.
- [ ] Yes. <!-- Add document PR link here. eg:
https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR
should merge into -->

+++ be/src/vec/common/sort/sorter.h
@@ -38,6 +38,7 @@
 #include "vec/utils/util.hpp"
 
 namespace doris {
+#include "common/compile_check_begin.h"
 class ObjectPool;
 class RowDescriptor;
 } // namespace doris
@@ -100,7 +101,7 @@ private:
 
 class Sorter {
 public:
-    Sorter(VSortExecExprs& vsort_exec_exprs, int limit, int64_t offset, ObjectPool* pool,
+    Sorter(VSortExecExprs& vsort_exec_exprs, int64_t limit, int64_t offset, ObjectPool* pool,
            std::vector<bool>& is_asc_order, std::vector<bool>& nulls_first)
             : _vsort_exec_exprs(vsort_exec_exprs),
               _limit(limit),
@@ -145,7 +146,7 @@ protected:
     bool _enable_spill = false;
     SortDescription _sort_description;
     VSortExecExprs& _vsort_exec_exprs;
-    int _limit;
+    int64_t _limit;
     int64_t _offset;
     ObjectPool* _pool = nullptr;
     std::vector<bool>& _is_asc_order;
@@ -163,7 +164,7 @@ class FullSorter final : public Sorter {
     ENABLE_FACTORY_CREATOR(FullSorter);
 
 public:
-    FullSorter(VSortExecExprs& vsort_exec_exprs, int limit, int64_t offset, ObjectPool* pool,
+    FullSorter(VSortExecExprs& vsort_exec_exprs, int64_t limit, int64_t offset, ObjectPool* pool,
                std::vector<bool>& is_asc_order, std::vector<bool>& nulls_first,
                const RowDescriptor& row_desc, RuntimeState* state, RuntimeProfile* profile);
 
@@ -195,4 +196,5 @@ private:
     static constexpr size_t INITIAL_BUFFERED_BLOCK_BYTES = 64 * 1024 * 1024;
 };
 
+#include "common/compile_check_end.h"
 } // namespace doris::vectorized
